requirement.txt
loguru~=0.5.3
requests~=2.26.0
pyTelegramBotAPI==3.8.3
telebot-calendar==1.2
python-dotenv==0.19.0